refactor: atualizar o local do loadingTimeout
