+++ frontend/src/features/freelancers/pages/PesquisaFreelancer.tsx
@@ -15,7 +15,7 @@ import { freelancerPaths } from "@/features/freelancers/routes/freelancerPaths";
 import Card from "@/shared/components/ui/card/Card";
 import clsx from "clsx";
 
-import { LOADING_TIMEOUT } from "@/app/loadingTimeout";
+import { LOADING_TIMEOUT } from "@/loadingTimeout";
 
 export function PesquisaFreelancer() {
   const navigate = useNavigate();
